Add download plot svg feature
This adds a button in the legend that takes the plot data a formats it
in a way to be used in a svg file. This works by taking the svg data
already used to display the plot, converting to base64 and directing the
user to a data: URI. This works all on the front end to make sure there
isn't any network needed.