Fixed "no operator" value/variable equations
Before if you gave jsut a calue or jsut a variable, the equation parser
would not give an error, but would just spit out the string of the value
given. A "no operator" operator was added that just returns it's only
operand. Had to override part of the equation toString to accomodate
this and make the resulting string not too confusing.